requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2023.7.6
PyNaCl>=1.5.0
python-dotenv>=1.0.0
spotipy>=2.23.0
aiofiles>=23.2.1
asyncio-throttle>=1.0.2
aiohttp>=3.8.0